@@ -1,4 +1,4 @@
-// +build integration
+//go:build integration
 
 package client
 
@@ -9,7 +9,7 @@ import (
 )
 
 /*
-	Test ShouldDeploy for a multitude of conditions.
+Test ShouldDeploy for a multitude of conditions.
 */
 func TestAutoDeployLatestTag(t *testing.T) {
 	te := SetUpClientTest(t)
